Style the form
