+++ src/components/BookItem/BookItem.css
@@ -1,7 +1,3 @@
-.item-middle {
-  width: 70px;
-  height: 70px;
-}
 
 .item-card {
   display: flex;
