+++ src/components/BookItem/ItemEnd/ItemEnd.css
@@ -1,0 +1,17 @@
+.update-button {
+ 
+  color: #e2e0e0;
+  padding: 10px 10%;
+  border-radius: 2px;
+  border: none;
+}
+
+
+
+.light-grey {
+  color: #eeeeee;
+}
+
+.grey {
+  color: #c1c1c1;
+}
