+++ src/components/BookItem/ItemEnd/ItemEnd.js
@@ -2,9 +2,9 @@ import './ItemEnd.css';
 
 const ItemEnd = () => (
   <div className="item-end">
-    <p>CURRENT CHAPTER</p>
-    <p>CHAPTER 17</p>
-    <button type="button">UPDATE PROGRESS</button>
+    <p className="light-grey">CURRENT CHAPTER</p>
+    <p className="grey">CHAPTER 17</p>
+    <button className="update-button bg-blue" type="button">UPDATE PROGRESS</button>
   </div>
 );
