+++ src/components/BookItem/ItemMiddle/ItemMiddle.js
@@ -3,7 +3,10 @@ import './ItemMiddle.css';
 
 const ItemMiddle = () => (
   <div className="item-middle">
-    <CircularProgressbar value={66} text={`${66}%`} />
+    <div className="progress-bar">
+      <CircularProgressbar value={66} text={`${66}%`} />
+    </div>
+    <p className="grey">Completed</p>
   </div>
 );
 export default ItemMiddle;
