+++ src/pages/Books.css
@@ -8,6 +8,4 @@
   margin: 20px 0;
 }
 
-.book-form {
-  padding: 0 10px;
-}
+
